pull changes to remove in-tree pybind11, refactor to allow python and c++ to coexist in a single package
