+++ python/datasketches/__init__.py
@@ -1,0 +1,7 @@
+#from .streaming_heap import StreamingHeap
+from .count_sketch import CountSketch
+
+try:
+  from _datasketches import *
+except ImportError:
+  print("Import error!")
