--- python/src/__init__.py
@@ -1,2 +1,0 @@
-name = "datasketches"
-
